+++ src/template-executor.ts
@@ -1,39 +1,5 @@
-import fs from 'fs';
-import path from 'path';
+import { FileTemplate } from './file-template';
 const changeCase = import('change-case');
-class FileToSave {
-    public variables: Record<string, string>
-    constructor(
-        public readonly content: string[],
-        variables: Record<string, string>
-    ) {
-        this.variables = JSON.parse(JSON.stringify(variables));
-    }
-
-    updateVariables(variables: Record<string, string>) {
-        this.variables = JSON.parse(JSON.stringify(variables));
-    }
-
-    addLine(value: string) {
-        this.content.push(value);
-    }
-
-    get filename() {
-        return `${this.getVariable('filename')}.${this.getVariable('ext') }`;
-    }
-
-    get text() {
-        return this.content.join('\n');
-    }
-
-    get folder() {
-        return this.variables['folder'];
-    }
-
-    getVariable(name: string) {
-        return this.variables[name] || `%${name}`;
-    }
-}
 
 export class TemplateExecutor{
     private pipes: Record<string, (input: string) => Promise<string>> = {
@@ -42,27 +8,32 @@ export class TemplateExecutor{
         ['pascal-case']: async (input: string) => (await changeCase).pascalCase(input),
         ['constant-case']: async (input: string) => (await changeCase).constantCase(input),
         ['path-case']: async (input: string) => (await changeCase).pathCase(input),
+        ['snake-case']: async (input: string) => (await changeCase).snakeCase(input),
+        ['no-case']: async (input: string) => (await changeCase).noCase(input),
+        ['pascal-snake-case']: async (input: string) => (await changeCase).pascalSnakeCase(input),
+        ['dot-case']: async (input: string) => (await changeCase).dotCase(input),
     };
     targetPath: string = '';
-    file: FileToSave[] = [];
+    files: FileTemplate[] = [];
     constructor(private variables: Record<string, string>, private template: string) {}
     setTargetPath(targetPath: string) {
         this.targetPath = targetPath;
     }
-    async exec() {
+    async exec(onStateChange: (state: string) => void) {
         const lines = this.template.split('\n');
-        this.file = [];
+        this.files = [];
         let fileIndex = 0;
+        onStateChange('Parsing');
         for (let i = 0; i < lines.length; i++) {
-            if (!this.file[fileIndex]) {
-                this.file[fileIndex] = new FileToSave([], this.variables);
+            if (!this.files[fileIndex]) {
+                this.files[fileIndex] = new FileTemplate([], this.variables);
             }
             const line = lines[i].replaceAll('\r', '');
 
             const trimmedLine = line.trim().replaceAll('\t', '');
             if (trimmedLine.startsWith('#')) {
-                if (trimmedLine.startsWith('#split')) {
-                    this.file[fileIndex].updateVariables(this.variables);
+                if (trimmedLine.startsWith('#split') || trimmedLine.startsWith('#beginfile') || trimmedLine.startsWith('#endfile')) {
+                    this.files[fileIndex].updateVariables(this.variables);
                     fileIndex++;
                 } else {
                     let [name, expression] = line.split(':');
@@ -70,29 +41,22 @@ export class TemplateExecutor{
                     expression = expression.trim();
                     const varName = name.replace('#', '');
                     this.variables[varName] = await this.parseExpression(expression);
-                    this.file[fileIndex].updateVariables(this.variables);
+                    this.files[fileIndex].updateVariables(this.variables);
                 }
             } else {
-                this.file[fileIndex].addLine(this.parseLine(line));
+                this.files[fileIndex].addLine(this.parseLine(line));
             }
         }
-        return this.file;
+        onStateChange('Parsing complete');
+        return this.files;
     }
 
-    save() {
-        for (const f of this.file) {
-            this.saveFile(f);
+    save(onStateChange: (state: string) => void) {
+        for (const f of this.files) {
+            onStateChange(`Saving file ${f.filename}`);
+            f.save(this.targetPath);
         }
-    }
-
-    saveFile(file: FileToSave) {
-        const result = file.text;
-        const folder = path.join(this.targetPath, file.folder);
-        fs.mkdirSync(folder, { recursive: true });
-        if (result && file.content.some(Boolean)) {
-            const resultPath = path.join(this.targetPath, file.folder, `./${file.filename}`);
-            fs.writeFileSync(resultPath, result, 'utf8',);
-        }
+        onStateChange(`Saving complete`);
     }
 
     parseLine(text: string) {