+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "file-templates" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.5]
 
-- Initial release
+- Add cases no-case snake-case pascal-snake-case dot-case
+- Status bar
+- Refactoring classes